fix: orders to match epic functionality
